Extract type into its own module
