+++ hangman/lib/impl/game.ex
@@ -1,7 +1,7 @@
 defmodule Hangman.Impl.Game do
   @type t :: %__MODULE__{
           turns_left: integer(),
-          game_state: Hangman.state(),
+          game_state: Type.state(),
           letters: list(String.t()),
           used: MapSet.t(String.t())
         }
